Add files via upload
Created a main file and coded the method to utilize the model to predict. 
Just need to further fine-tune some parameters for my model and modify my dataset to include more training and validation data.
Was unable to commit any changes to the project due to some family issues alongside a sudden increase in workload.